refactor(accelerator): extract updater module, eliminate redundant network check
Move check_for_update() and perform_update() from main.rs to a shared
updater.rs module in the lib. This lets both the background loop (main.rs)
and the "Update Now" button (commands.rs) use the same code.

Before: respond_update_prompt re-checked the network from scratch (~40
lines of duplicated updater logic). The update shown in the prompt could
theoretically differ from the one downloaded.

After: check_for_update stores the Update object in PendingUpdate managed
state. respond_update_prompt takes it directly and calls perform_update.
Zero redundant network requests, no race condition.

New: src/updater.rs with check_for_update() and perform_update()
New: PendingUpdate type (Arc<Mutex<Option<Update>>>) in commands.rs
Removed: ~40 lines of duplicated updater logic from commands.rs
